# src/main/resources/api.yml
openapi: 3.1.1
info:
  title: BaroLab API
  version: 0.0.1
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /posts:
    post:
      operationId: createPost
      tags:
        - Posts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: getAllPosts
      summary: Get all Posts
      tags:
        - Posts
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /post/{postId}:
    get:
      operationId: getPostById
      parameters:
        - $ref: '#/components/parameters/postId'
      summary: Get Post by ID
      tags:
        - Posts
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post' 
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /mods:
    post:
      operationId: createMod
      tags:
        - Mods
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModPostRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModPost'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: getAllMods
      summary: Get all Mods
      tags:
        - Mods
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModPost'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mod/{externalId}:
    get:
      operationId: getModByExternalId
      parameters:
        - $ref: '#/components/parameters/externalId'
      summary: Get Mod by external ID
      tags:
        - Mods
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModPost'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mod/{externalId}/transition:
    post:
      operationId: transitionToMod
      security: []
      parameters:
        - $ref: '#/components/parameters/externalId'
      summary: Redirect to Steam Workshop and register click (cooldown 1h)
      tags:
        - Mods
      responses:
        302:
          description: Redirect to Steam Workshop
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mod/{externalId}/activate:
    put:
      operationId: activateMod
      parameters:
        - $ref: '#/components/parameters/externalId'
      tags:
        - Mods
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModPost'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /mod/{externalId}/block:
    put:
      operationId: blockMod
      parameters:
        - $ref: '#/components/parameters/externalId'
      tags:
        - Mods
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModPost'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    postId:
      name: postId
      in: path
      required: true
      schema:
        type: string 
        format: uuid
      description: Id of Post
...
    externalId:
      name: externalId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Steam Workshop external id
  schemas:
...
    Post:
      type: object
      required:
        - id
        - user_id
        - author_username
        - rating
        - status
        - title
        - content
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: 38ca31a9-a336-4087-ac97-d1ed40a2778e
          readOnly: true
        user_id:
          type: string
          format: uuid
          example: 38ca31a9-a336-4087-ac97-d1ed40a2778e
          readOnly: true
        author_username:
          type: string
          readOnly: true
        rating:
          type: integer
          example: 0
          readOnly: true
        my_vote:
          type: integer
          enum: [ -1, 1 ]
          example: 1
          readOnly: true
        status:
          type: string 
          enum: [ ACTIVE, BLOCKED ]
          example: ACTIVE
          readOnly: true
        title:
          type: string
          example: Post name
        content:
          type: string 
          example: content
        created_at:
          type: string
          format: date-time
          example: 2021-09-20T12:00:00Z
          readOnly: true
        updated_at:
          type: string
          format: date-time
          example: 2021-09-20T13:41:22Z
          readOnly: true
    CreatePostRequest:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          example: Post name
        content:
          type: string
          example: content
    ModPost:
      type: object
      required:
        - id
        - user_id
        - author_username
        - rating
        - status
        - title
        - content
        - created_at
        - updated_at
        - external_id
        - popularity
      properties:
        id:
          type: string
          format: uuid
          example: 38ca31a9-a336-4087-ac97-d1ed40a2778e
          readOnly: true
        user_id:
          type: string
          format: uuid
          example: 38ca31a9-a336-4087-ac97-d1ed40a2778e
          readOnly: true
        author_username:
          type: string
          readOnly: true
        rating:
          type: integer
          example: 0
          readOnly: true
        my_vote:
          type: integer
          enum: [ -1, 1 ]
          example: 1
          readOnly: true
        status:
          type: string
          enum: [ ACTIVE, BLOCKED ]
          example: ACTIVE
          readOnly: true
        title:
          type: string
          example: Mod name
        content:
          type: string
          example: content
        created_at:
          type: string
          format: date-time
          example: 2021-09-20T12:00:00Z
          readOnly: true
        updated_at:
          type: string
          format: date-time
          example: 2021-09-20T13:41:22Z
          readOnly: true
        external_id:
          type: integer
          format: int64
          example: 3520553488
          readOnly: true
        popularity:
          type: integer
          example: 0
          readOnly: true
    CreateModPostRequest:
      type: object
      required:
        - title
        - content
        - external_url
      properties:
        title:
          type: string
          example: Mod name
        content:
          type: string
          example: content
        external_url:
          type: string
          example: https://steamcommunity.com/sharedfiles/filedetails/?id=3520553488
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Error message.